Extract CommentingIdeaDisabled
